turn propose_parachains into validator_manager
